+++ ProphetsWay.Example.DataAccess/Entities/Job.cs
@@ -1,6 +1,6 @@
 ﻿namespace ProphetsWay.Example.DataAccess.Entities
 {
-	public class Job : BaseEntity
+	public class Job : BaseIntEntity
 	{
 		public string Name { get; set; }