+++ ProphetsWay.Example.Database/dbo/Tables/Resources.sql
@@ -1,0 +1,6 @@
+﻿CREATE TABLE [dbo].[Resources]
+(
+	[Id] UNIQUEIDENTIFIER NOT NULL PRIMARY KEY, 
+	[Name] VARCHAR(50) NOT NULL, 
+	[Description] VARCHAR(MAX) NULL
+)